+++ reptar/__init__.py
@@ -3,16 +3,7 @@ __author__ = """Alex M. Maldonado"""
 from . import _version
 __version__ = _version.get_versions()['version']
 
-from . import utils
-from . import extractors
-from . import parsers
 from .creator import creator
 from .reptar_file import File
-from . import sampler
-from . import writers
-from . import fragment
 
-__all__ = [
-    'utils', 'extractors', 'parsers', 'creator', 'File', 'sampler', 'writers',
-    'fragment'
-]
+__all__ = ['creator', 'File']