+++ reptar/writers/xyz_gap.py
@@ -32,27 +32,25 @@ def write_xyz_gap(
     ----------
     xyz_path : :obj:`str`
         Path to the XYZ file to write.
-    lattice : :obj:`numpy.ndarray`
+    lattice : :obj:`numpy.ndarray`, ndim: ``2``
         The three cartesian lattice vectors describing the periodic cell (in
         Angstroms).
         
         All structures need a this lattice even structures are not supposed to
         be periodic. Just use a lattice vector that is larger than twice the
         cutoff of any potential you plan to create or use.
-    Z : :obj:`numpy.ndarray`
+    Z : :obj:`numpy.ndarray`, ndim: ``1``
         Atomic numbers of all atoms in the system.
-    R : :obj:`numpy.ndarray`
+    R : :obj:`numpy.ndarray`, ndim: ``3``
         Cartesian coordinates of all structures in the same order as ``Z``.
-    E : :obj:`numpy.ndarray`
+    E : :obj:`numpy.ndarray`, ndim: ``1``
         Energy of all structures in ``R`` in units of eV.
-    F : :obj:`numpy.ndarray`, optional
-        Atomic forces of all structures in ``R`` in units of eV/A. Defaults to
-        ``None``.
-    lattice_precision : :obj:`int`, optional
-        Number of decimal points to print for lattice dimensions. Default is
-        ``3``.
-    data_precision : :obj:`int`, optional
-        Number of decimal points for printing array data. Default is ``13``.
+    F : :obj:`numpy.ndarray`, ndim: ``3``, default: ``None``
+        Atomic forces of all structures in ``R`` in units of eV/A.
+    lattice_precision : :obj:`int`, default: ``3``
+        Number of decimal points to print for lattice dimensions.
+    data_precision : :obj:`int`, default: ``10``
+        Number of decimal points for printing array data.
 
     """
     lat_str = np.array2string(